Adjust the author dropdown.
